Use a separate port for Akka; upgrade Protobuf 2.3.0 -> 2.4.1
Akka can't bind to the same port as Scala actors, so this commit adds
a new property spark.master.akkaPort that defaults to 50502.

Akka 1.2 is incompatible with Protobuf 2.3.0 and needs 2.4.1 instead,
as discussed in
http://groups.google.com/group/akka-user/msg/80cee7566f0b5741. This
commit moves to Protobuf 2.4.1. This may cause problems with other
parts of Spark; I haven't tested it.