Report checksums of RDD splits
Checksumming the values in RDDs can make it possible to detect subtle
bugs caused by indeterminism in user code. This commit intercepts the
output of the RDD compute() method, calculates checksums of the
computed values, and reports the checksums back to the master.

It does this by wrapping the output of compute(), an Iterator, in a
ChecksummingIterator. This class maintains a running checksum of the
values that come out of the underlying Iterator and reports the
cumulative checksum to the master when the iterator is exhausted.