Standardize event log format; allow setting RDD context after deserialization
This commit introduces a standard format for the event log: a sequence
of serialized EventReporterMessages.

It also introduces a new method, RDD.setContext, for setting the
SparkContext of an RDD and its dependencies after deserializing it
from the event log. Unfortunately, this means that RDDs are no longer
immutable, so I'm looking for a better solution. (That's why this
commit is in a separate branch.)

A known bug is that ParallelCollections don't get serialized properly
because their data is currently marked @transient.